Changed bit handling.

+++ include/huffman.h
@@ -6,7 +6,7 @@
 
 
 
-uint16_t* huffmanTable(size_t* code_lengths, size_t code_count, size_t* table_size, size_t* max_code_length);
+unsigned int* huffmanTable(size_t* code_lengths, size_t code_count, size_t* table_size, size_t* max_code_length);
